Updated clocks to remove all warnings.
