+++ clocks/countdown.py
@@ -1,5 +1,3 @@
-import time
-
 from clocks.clock import Clock
 
 
@@ -7,7 +5,7 @@ class Countdown(Clock):
     def __init__(self, player_count=0, player_names=None, starting_clock=5):
         super().__init__(player_count=player_count, player_names=player_names)
 
-        self.clock = [int(starting_clock * 60) for i in range(self.player_count+1)]
+        self.clock = [int(starting_clock * 60) for _ in range(self.player_count+1)]
 
     def next_player(self):
         # Get turn length, and then add it to the clock
